detele python related and rename some files
