--- tests/benchmarks/micro_benchmarks/test_cublas_functions.py
@@ -1,46 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-
-"""Tests for cublas-functions benchmark."""
-
-import numbers
-import json
-
-from tests.helper import decorator
-from superbench.benchmarks import BenchmarkRegistry, BenchmarkType, ReturnCode
-
-
-@decorator.cuda_test
-def test_cublas_functions():
-    """Test cublas-function benchmark."""
-    context = BenchmarkRegistry.create_benchmark_context(
-        'cublas-functions', parameters='--num_warmup 10 --num_steps 10 --num_in_step 100'
-    )
-
-    assert (BenchmarkRegistry.is_benchmark_context_valid(context))
-
-    benchmark = BenchmarkRegistry.launch_benchmark(context)
-
-    # Check basic information.
-    assert (benchmark)
-    assert (benchmark.name == 'cublas-functions')
-    assert (benchmark.type == BenchmarkType.MICRO)
-
-    # Check parameters specified in BenchmarkContext.
-    assert (benchmark._args.num_warmup == 10)
-    assert (benchmark._args.num_steps == 10)
-    assert (benchmark._args.num_in_step == 100)
-
-    # Check results and metrics.
-    assert (benchmark.run_count == 1)
-    assert (benchmark.return_code == ReturnCode.SUCCESS)
-    assert ('raw_output_0' in benchmark.raw_data)
-    assert (len(benchmark.raw_data['raw_output_0']) == 1)
-    assert (isinstance(benchmark.raw_data['raw_output_0'][0], str))
-    config_path = '/opt/superbench/superbench/benchmarks/micro_benchmarks/cublas_function/para_info.json'
-    with open(config_path, 'r') as load_f:
-        load_json = json.load(load_f)
-        assert (len(load_json) == len(benchmark.result))
-    for metric in list(benchmark.result.keys()):
-        assert (len(benchmark.result[metric]) == 1)
-        assert (isinstance(benchmark.result[metric][0], numbers.Number))
